% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fef_SSCO_Categories.R
\name{fef_SSCO_Categories}
\alias{fef_SSCO_Categories}
\title{SSCO Incident Categories, derived from short description}
\usage{
fef_SSCO_Categories(df_object, col_string)
}
\arguments{
\item{col_string}{col name in string format}

\item{df}{df object}
}
\description{
SSCO Incident Categories, derived from short description
}
\examples{
test
}
\keyword{calculated}
\keyword{fields,}
\keyword{regex}
