% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau_SSCO_Categories.R
\name{tableau_SSCO_Categories}
\alias{tableau_SSCO_Categories}
\title{SSCO Incident Categories, derived from short description.}
\usage{
tableau_SSCO_Categories(col_string)
}
\arguments{
\item{col_string}{col name in string format}
}
\description{
SSCO Incident Categories, derived from short description.
}
\examples{
test
}
\keyword{calculated}
\keyword{fields,}
\keyword{regex}
